requirements.txt
cffi==1.15.1
importlib-metadata==6.1.0
jaraco.classes==3.2.3
keyring==23.13.1
more-itertools==9.1.0
mysql==0.0.3
mysqlclient==2.1.1
pycparser==2.21
PyNaCl==1.5.0
pywin32-ctypes==0.2.0
zipp==3.15.0
